Put inner classes in separate files
