+++ android5.0/Camera2Basic/Camera2Basic/Camera2BasicFragment.cs
@@ -1,6 +1,4 @@
-﻿
-
-using System;
+﻿using System;
 using System.Collections.Generic;
 using Android;
 using Android.App;
@@ -16,15 +14,13 @@ using Android.Hardware.Camera2.Params;
 using Android.Media;
 using Android.Support.V13.App;
 using Android.Support.V4.Content;
+using Camera2Basic.Listeners;
 using Java.IO;
 using Java.Lang;
-using Java.Nio;
 using Java.Util;
 using Java.Util.Concurrent;
 using Boolean = Java.Lang.Boolean;
-using CameraError = Android.Hardware.Camera2.CameraError;
 using Math = Java.Lang.Math;
-using Object = Java.Lang.Object;
 using Orientation = Android.Content.Res.Orientation;
 
 namespace Camera2Basic
@@ -32,26 +28,26 @@ namespace Camera2Basic
     public class Camera2BasicFragment : Fragment, View.IOnClickListener, FragmentCompat.IOnRequestPermissionsResultCallback
     {
         private static readonly SparseIntArray ORIENTATIONS = new SparseIntArray();
-        private static readonly int REQUEST_CAMERA_PERMISSION = 1;
+        public static readonly int REQUEST_CAMERA_PERMISSION = 1;
         private static readonly string FRAGMENT_DIALOG = "dialog";
 
         // Tag for the {@link Log}.
         private static readonly string TAG = "Camera2BasicFragment";
 
         // Camera state: Showing camera preview.
-        private const int STATE_PREVIEW = 0;
+        public const int STATE_PREVIEW = 0;
 
         // Camera state: Waiting for the focus to be locked.
-        private const int STATE_WAITING_LOCK = 1;
+        public const int STATE_WAITING_LOCK = 1;
 
         // Camera state: Waiting for the exposure to be precapture state.
-        private const int STATE_WAITING_PRECAPTURE = 2;
+        public const int STATE_WAITING_PRECAPTURE = 2;
 
         //Camera state: Waiting for the exposure state to be something other than precapture.
-        private const int STATE_WAITING_NON_PRECAPTURE = 3;
+        public const int STATE_WAITING_NON_PRECAPTURE = 3;
 
         // Camera state: Picture was taken.
-        private const int STATE_PICTURE_TAKEN = 4;
+        public const int STATE_PICTURE_TAKEN = 4;
 
         // Max preview width that is guaranteed by Camera2 API
         private static readonly int MAX_PREVIEW_WIDTH = 1920;
@@ -61,36 +57,6 @@ namespace Camera2Basic
 
         // TextureView.ISurfaceTextureListener handles several lifecycle events on a TextureView
         private Camera2BasicSurfaceTextureListener mSurfaceTextureListener;
-        private class Camera2BasicSurfaceTextureListener : Java.Lang.Object, TextureView.ISurfaceTextureListener
-        {
-            public Camera2BasicFragment Owner { get; set; }
-
-            public Camera2BasicSurfaceTextureListener(Camera2BasicFragment owner)
-            {
-                Owner = owner;
-            }
-
-            public void OnSurfaceTextureAvailable(Android.Graphics.SurfaceTexture surface, int width, int height)
-            {
-                Owner.OpenCamera(width, height);
-            }
-
-            public bool OnSurfaceTextureDestroyed(Android.Graphics.SurfaceTexture surface)
-            {
-                return true;
-            }
-
-            public void OnSurfaceTextureSizeChanged(Android.Graphics.SurfaceTexture surface, int width, int height)
-            {
-                Owner.ConfigureTransform(width, height);
-            }
-
-            public void OnSurfaceTextureUpdated(Android.Graphics.SurfaceTexture surface)
-            {
-
-            }
-        }
-
 
         // ID of the current {@link CameraDevice}.
         private string mCameraId;
@@ -99,89 +65,44 @@ namespace Camera2Basic
         private AutoFitTextureView mTextureView;
 
         // A {@link CameraCaptureSession } for camera preview.
-        private CameraCaptureSession mCaptureSession;
+        public CameraCaptureSession mCaptureSession;
 
         // A reference to the opened CameraDevice
-        private CameraDevice mCameraDevice;
+        public CameraDevice mCameraDevice;
 
         // The size of the camera preview
         private Size mPreviewSize;
 
         // CameraDevice.StateListener is called when a CameraDevice changes its state
         private CameraStateListener mStateCallback;
-        private class CameraStateListener : CameraDevice.StateCallback
-        {
-            public Camera2BasicFragment owner;
-            public override void OnOpened(CameraDevice cameraDevice)
-            {
-                // This method is called when the camera is opened.  We start camera preview here.
-                owner.mCameraOpenCloseLock.Release();
-                owner.mCameraDevice = cameraDevice;
-                owner.CreateCameraPreviewSession();
-            }
-
-            public override void OnDisconnected(CameraDevice cameraDevice)
-            {
-                owner.mCameraOpenCloseLock.Release();
-                cameraDevice.Close();
-                owner.mCameraDevice = null;
-            }
-
-            public override void OnError(CameraDevice cameraDevice, CameraError error)
-            {
-                owner.mCameraOpenCloseLock.Release();
-                cameraDevice.Close();
-                owner.mCameraDevice = null;
-                if (owner == null)
-                    return;
-                Activity activity = owner.Activity;
-                if (activity != null)
-                {
-                    activity.Finish();
-                }
-
-            }
-        }
 
         // An additional thread for running tasks that shouldn't block the UI.
         private HandlerThread mBackgroundThread;
 
-
         // A {@link Handler} for running tasks in the background.
-        private Handler mBackgroundHandler;
-
+        public Handler mBackgroundHandler;
 
         // An {@link ImageReader} that handles still image capture.
         private ImageReader mImageReader;
 
         // This is the output file for our picture.
-        private File mFile;
+        public File mFile;
 
         // This a callback object for the {@link ImageReader}. "onImageAvailable" will be called when a
         // still image is ready to be saved.
         private ImageAvailableListener mOnImageAvailableListener;
-        private class ImageAvailableListener : Java.Lang.Object, ImageReader.IOnImageAvailableListener
-        {
-            public File File { get; set; }
-            public Camera2BasicFragment Owner { get; set; }
-            public void OnImageAvailable(ImageReader reader)
-            {
-                Owner.mBackgroundHandler.Post(new ImageSaver(reader.AcquireNextImage(), File));
-            }
-        }
-
 
         //{@link CaptureRequest.Builder} for the camera preview
-        private CaptureRequest.Builder mPreviewRequestBuilder;
+        public CaptureRequest.Builder mPreviewRequestBuilder;
 
         // {@link CaptureRequest} generated by {@link #mPreviewRequestBuilder}
-        private CaptureRequest mPreviewRequest;
+        public CaptureRequest mPreviewRequest;
 
         // The current state of camera state for taking pictures.
-        private int mState = STATE_PREVIEW;
+        public int mState = STATE_PREVIEW;
 
         // A {@link Semaphore} to prevent the app from exiting before closing the camera.
-        private Semaphore mCameraOpenCloseLock = new Semaphore(1);
+        public Semaphore mCameraOpenCloseLock = new Semaphore(1);
 
         // Whether the current camera device supports Flash or not.
         private bool mFlashSupported;
@@ -190,80 +111,10 @@ namespace Camera2Basic
         private int mSensorOrientation;
 
         // A {@link CameraCaptureSession.CaptureCallback} that handles events related to JPEG capture.
-        private CameraCaptureListener mCaptureCallback;
-        private class CameraCaptureListener : CameraCaptureSession.CaptureCallback
-        {
-            public Camera2BasicFragment Owner { get; set; }
-            public File File { get; set; }
-            public override void OnCaptureCompleted(CameraCaptureSession session, CaptureRequest request, TotalCaptureResult result)
-            {
-                Process(result);
-            }
-
-            public override void OnCaptureProgressed(CameraCaptureSession session, CaptureRequest request, CaptureResult partialResult)
-            {
-                Process(partialResult);
-            }
-
-            private void Process(CaptureResult result)
-            {
-                switch (Owner.mState)
-                {
-                    case STATE_WAITING_LOCK:
-                        {
-                            Integer afState = (Integer)result.Get(CaptureResult.ControlAfState);
-                            if (afState == null)
-                            {
-                                Owner.CaptureStillPicture();
-                            }
-
-                            else if ((((int)ControlAFState.FocusedLocked) == afState.IntValue()) ||
-                                       (((int)ControlAFState.NotFocusedLocked) == afState.IntValue()))
-                            {
-                                // ControlAeState can be null on some devices
-                                Integer aeState = (Integer)result.Get(CaptureResult.ControlAeState);
-                                if (aeState == null ||
-                                        aeState.IntValue() == ((int)ControlAEState.Converged))
-                                {
-                                    Owner.mState = STATE_PICTURE_TAKEN;
-                                    Owner.CaptureStillPicture();
-                                }
-                                else
-                                {
-                                    Owner.RunPrecaptureSequence();
-                                }
-                            }
-                            break;
-                        }
-                    case STATE_WAITING_PRECAPTURE:
-                        {
-                            // ControlAeState can be null on some devices
-                            Integer aeState = (Integer)result.Get(CaptureResult.ControlAeState);
-                            if (aeState == null ||
-                                    aeState.IntValue() == ((int)ControlAEState.Precapture) ||
-                                    aeState.IntValue() == ((int)ControlAEState.FlashRequired))
-                            {
-                                Owner.mState = STATE_WAITING_NON_PRECAPTURE;
-                            }
-                            break;
-                        }
-                    case STATE_WAITING_NON_PRECAPTURE:
-                        {
-                            // ControlAeState can be null on some devices
-                            Integer aeState = (Integer)result.Get(CaptureResult.ControlAeState);
-                            if (aeState == null || aeState.IntValue() != ((int)ControlAEState.Precapture))
-                            {
-                                Owner.mState = STATE_PICTURE_TAKEN;
-                                Owner.CaptureStillPicture();
-                            }
-                            break;
-                        }
-                }
-            }
-        }
+        public CameraCaptureListener mCaptureCallback;
 
         // Shows a {@link Toast} on the UI thread.
-        private void ShowToast(string text)
+        public void ShowToast(string text)
         {
             if (Activity != null)
             {
@@ -320,11 +171,11 @@ namespace Camera2Basic
             // largest of those not big enough.
             if (bigEnough.Count > 0)
             {
-                return (Size) Collections.Min(bigEnough, new CompareSizesByArea());
+                return (Size)Collections.Min(bigEnough, new CompareSizesByArea());
             }
             else if (notBigEnough.Count > 0)
             {
-                return (Size) Collections.Max(notBigEnough, new CompareSizesByArea());
+                return (Size)Collections.Max(notBigEnough, new CompareSizesByArea());
             }
             else
             {
@@ -447,7 +298,7 @@ namespace Camera2Basic
                     }
 
                     // For still image captures, we use the largest available size.
-                    Size largest = (Size) Collections.Max(Arrays.AsList(map.GetOutputSizes((int)ImageFormatType.Jpeg)),
+                    Size largest = (Size)Collections.Max(Arrays.AsList(map.GetOutputSizes((int)ImageFormatType.Jpeg)),
                         new CompareSizesByArea());
                     mImageReader = ImageReader.NewInstance(largest.Width, largest.Height, ImageFormatType.Jpeg, /*maxImages*/2);
                     mImageReader.SetOnImageAvailableListener(mOnImageAvailableListener, mBackgroundHandler);
@@ -550,7 +401,7 @@ namespace Camera2Basic
         }
 
         // Opens the camera specified by {@link Camera2BasicFragment#mCameraId}.
-        private void OpenCamera(int width, int height)
+        public void OpenCamera(int width, int height)
         {
             if (ContextCompat.CheckSelfPermission(Activity, Manifest.Permission.Camera) != Permission.Granted)
             {
@@ -636,7 +487,7 @@ namespace Camera2Basic
         }
 
         // Creates a new {@link CameraCaptureSession} for camera preview.
-        private void CreateCameraPreviewSession()
+        public void CreateCameraPreviewSession()
         {
             try
             {
@@ -669,50 +520,6 @@ namespace Camera2Basic
             }
         }
 
-        private class CameraCaptureSessionCallback : CameraCaptureSession.StateCallback
-        {
-
-            private Camera2BasicFragment owner;
-
-            public CameraCaptureSessionCallback(Camera2BasicFragment owner)
-            {
-                this.owner = owner;
-            }
-
-            public override void OnConfigureFailed(CameraCaptureSession session)
-            {
-                owner.ShowToast("Failed");
-            }
-
-            public override void OnConfigured(CameraCaptureSession session)
-            {
-                // The camera is already closed
-                if (null == owner.mCameraDevice)
-                {
-                    return;
-                }
-
-                // When the session is ready, we start displaying the preview.
-                owner.mCaptureSession = session;
-                try
-                {
-                    // Auto focus should be continuous for camera preview.
-                    owner.mPreviewRequestBuilder.Set(CaptureRequest.ControlAfMode, (int)ControlAFMode.ContinuousPicture);
-                    // Flash is automatically enabled when necessary.
-                    owner.SetAutoFlash(owner.mPreviewRequestBuilder);
-
-                    // Finally, we start displaying the camera preview.
-                    owner.mPreviewRequest = owner.mPreviewRequestBuilder.Build();
-                    owner.mCaptureSession.SetRepeatingRequest(owner.mPreviewRequest,
-                            owner.mCaptureCallback, owner.mBackgroundHandler);
-                }
-                catch (CameraAccessException e)
-                {
-                    e.PrintStackTrace();
-                }
-            }
-        }
-
         public static T Cast<T>(Java.Lang.Object obj) where T : class
         {
             var propertyInfo = obj.GetType().GetProperty("Instance");
@@ -724,7 +531,7 @@ namespace Camera2Basic
         // This method should be called after the camera preview size is determined in
         // setUpCameraOutputs and also the size of `mTextureView` is fixed.
 
-        private void ConfigureTransform(int viewWidth, int viewHeight)
+        public void ConfigureTransform(int viewWidth, int viewHeight)
         {
             Activity activity = Activity;
             if (null == mTextureView || null == mPreviewSize || null == activity)
@@ -779,7 +586,7 @@ namespace Camera2Basic
 
         // Run the precapture sequence for capturing a still image. This method should be called when
         // we get a response in {@link #mCaptureCallback} from {@link #lockFocus()}.
-        private void RunPrecaptureSequence()
+        public void RunPrecaptureSequence()
         {
             try
             {
@@ -797,7 +604,7 @@ namespace Camera2Basic
 
         // Capture a still picture. This method should be called when we get a response in
         // {@link #mCaptureCallback} from both {@link #lockFocus()}.
-        private void CaptureStillPicture()
+        public void CaptureStillPicture()
         {
             try
             {
@@ -827,23 +634,6 @@ namespace Camera2Basic
             }
         }
 
-        private class CameraCaptureStillPictureSessionCallback : CameraCaptureSession.CaptureCallback
-        {
-            private Camera2BasicFragment owner;
-
-            public CameraCaptureStillPictureSessionCallback(Camera2BasicFragment owner)
-            {
-                this.owner = owner;
-            }
-
-            public override void OnCaptureCompleted(CameraCaptureSession session, CaptureRequest request, TotalCaptureResult result)
-            {
-                owner.ShowToast("Saved: " + owner.mFile);
-                Log.Debug(TAG, owner.mFile.ToString());
-                owner.UnlockFocus();
-            }
-        }
-
         // Retrieves the JPEG orientation from the specified screen rotation.
         private int GetOrientation(int rotation)
         {
@@ -856,7 +646,7 @@ namespace Camera2Basic
 
         // Unlock the focus. This method should be called when still image capture sequence is
         // finished.
-        private void UnlockFocus()
+        public void UnlockFocus()
         {
             try
             {
@@ -897,131 +687,13 @@ namespace Camera2Basic
             }
         }
 
-        private void SetAutoFlash(CaptureRequest.Builder requestBuilder)
+        public void SetAutoFlash(CaptureRequest.Builder requestBuilder)
         {
             if (mFlashSupported)
             {
                 requestBuilder.Set(CaptureRequest.ControlAeMode, (int)ControlAEMode.OnAutoFlash);
             }
         }
-
-        // Saves a JPEG {@link Image} into the specified {@link File}.
-        private class ImageSaver : Java.Lang.Object, IRunnable
-        {
-            // The JPEG image
-            private Image mImage;
-
-            // The file we save the image into.
-            private File mFile;
-
-            public ImageSaver(Image image, File file)
-            {
-                mImage = image;
-                mFile = file;
-            }
-
-            public void Run()
-            {
-                ByteBuffer buffer = mImage.GetPlanes()[0].Buffer;
-                byte[] bytes = new byte[buffer.Remaining()];
-                buffer.Get(bytes);
-                using (var output = new FileOutputStream(mFile))
-                {
-                    try
-                    {
-                        output.Write(bytes);
-                    }
-                    catch (IOException e)
-                    {
-                        e.PrintStackTrace();
-                    }
-                    finally
-                    {
-                        mImage.Close();
-                    }
-                }
-            }
-        }
-
-        // Compares two {@code Size}s based on their areas.
-        public class CompareSizesByArea : Java.Lang.Object, IComparator
-        {
-            public int Compare(Object lhs, Object rhs)
-            {
-                var lhsSize = (Size)lhs;
-                var rhsSize = (Size)rhs;
-                // We cast here to ensure the multiplications won't overflow
-                return Long.Signum((long)lhsSize.Width * lhsSize.Height - (long)rhsSize.Width * rhsSize.Height);
-            }
-        }
-
-        // Shows an error message dialog.
-        public class ErrorDialog : DialogFragment
-        {
-            private static readonly string ARG_MESSAGE = "message";
-            private static Activity mActivity;
-
-            private class PositiveListener : Java.Lang.Object, IDialogInterfaceOnClickListener
-            {
-                public void OnClick(IDialogInterface dialog, int which)
-                {
-                    mActivity.Finish();
-                }
-            }
-
-            public static ErrorDialog NewInstance(string message)
-            {
-                var args = new Bundle();
-                args.PutString(ARG_MESSAGE, message);
-                return new ErrorDialog { Arguments = args };
-            }
-
-            public override Dialog OnCreateDialog(Bundle savedInstanceState)
-            {
-                mActivity = Activity;
-                return new AlertDialog.Builder(mActivity)
-                    .SetMessage(Arguments.GetString(ARG_MESSAGE))
-                    .SetPositiveButton(Android.Resource.String.Ok, new PositiveListener())
-                    .Create();
-            }
-        }
-
-        // Shows OK/Cancel confirmation dialog about camera permission.
-        public class ConfirmationDialog : DialogFragment
-        {
-            private static Fragment mParent;
-            private class PositiveListener : Java.Lang.Object, IDialogInterfaceOnClickListener
-            {
-                public void OnClick(IDialogInterface dialog, int which)
-                {
-                    FragmentCompat.RequestPermissions(mParent,
-                                    new string[] { Manifest.Permission.Camera }, REQUEST_CAMERA_PERMISSION);
-                }
-            }
-
-            private class NegativeListener : Java.Lang.Object, IDialogInterfaceOnClickListener
-            {
-                public void OnClick(IDialogInterface dialog, int which)
-                {
-                    Activity activity = mParent.Activity;
-                    if (activity != null)
-                    {
-                        activity.Finish();
-                    }
-                }
-            }
-
-            public override Dialog OnCreateDialog(Bundle savedInstanceState)
-            {
-                mParent = ParentFragment;
-                return new AlertDialog.Builder(Activity)
-                    .SetMessage(Resource.String.request_permission)
-                    .SetPositiveButton(Android.Resource.String.Ok, new PositiveListener())
-                    .SetNegativeButton(Android.Resource.String.Cancel, new NegativeListener())
-                    .Create();
-            }
-        }
-
     }
 }
