+++ android5.0/Camera2Basic/Camera2Basic/ConfirmationDialog.cs
@@ -1,0 +1,44 @@
+
+using Android;
+using Android.App;
+using Android.Content;
+using Android.OS;
+using Android.Support.V13.App;
+
+namespace Camera2Basic
+{
+    public class ConfirmationDialog : DialogFragment
+    {
+        private static Fragment mParent;
+        private class PositiveListener : Java.Lang.Object, IDialogInterfaceOnClickListener
+        {
+            public void OnClick(IDialogInterface dialog, int which)
+            {
+                FragmentCompat.RequestPermissions(mParent,
+                                new string[] { Manifest.Permission.Camera }, Camera2BasicFragment.REQUEST_CAMERA_PERMISSION);
+            }
+        }
+
+        private class NegativeListener : Java.Lang.Object, IDialogInterfaceOnClickListener
+        {
+            public void OnClick(IDialogInterface dialog, int which)
+            {
+                Activity activity = mParent.Activity;
+                if (activity != null)
+                {
+                    activity.Finish();
+                }
+            }
+        }
+
+        public override Dialog OnCreateDialog(Bundle savedInstanceState)
+        {
+            mParent = ParentFragment;
+            return new AlertDialog.Builder(Activity)
+                .SetMessage(Resource.String.request_permission)
+                .SetPositiveButton(Android.Resource.String.Ok, new PositiveListener())
+                .SetNegativeButton(Android.Resource.String.Cancel, new NegativeListener())
+                .Create();
+        }
+    }
+}
