+++ android5.0/Camera2Basic/Camera2Basic/Listeners/Camera2BasicSurfaceTextureListener.cs
@@ -1,0 +1,35 @@
+
+using Android.Views;
+
+namespace Camera2Basic.Listeners
+{
+    public class Camera2BasicSurfaceTextureListener : Java.Lang.Object, TextureView.ISurfaceTextureListener
+    {
+        public Camera2BasicFragment Owner { get; set; }
+
+        public Camera2BasicSurfaceTextureListener(Camera2BasicFragment owner)
+        {
+            Owner = owner;
+        }
+
+        public void OnSurfaceTextureAvailable(Android.Graphics.SurfaceTexture surface, int width, int height)
+        {
+            Owner.OpenCamera(width, height);
+        }
+
+        public bool OnSurfaceTextureDestroyed(Android.Graphics.SurfaceTexture surface)
+        {
+            return true;
+        }
+
+        public void OnSurfaceTextureSizeChanged(Android.Graphics.SurfaceTexture surface, int width, int height)
+        {
+            Owner.ConfigureTransform(width, height);
+        }
+
+        public void OnSurfaceTextureUpdated(Android.Graphics.SurfaceTexture surface)
+        {
+
+        }
+    }
+}
