+++ android5.0/Camera2Basic/Camera2Basic/Listeners/CameraStateListener.cs
@@ -1,0 +1,40 @@
+
+using Android.App;
+using Android.Hardware.Camera2;
+
+namespace Camera2Basic.Listeners
+{
+    public class CameraStateListener : CameraDevice.StateCallback
+    {
+        public Camera2BasicFragment owner;
+        public override void OnOpened(CameraDevice cameraDevice)
+        {
+            // This method is called when the camera is opened.  We start camera preview here.
+            owner.mCameraOpenCloseLock.Release();
+            owner.mCameraDevice = cameraDevice;
+            owner.CreateCameraPreviewSession();
+        }
+
+        public override void OnDisconnected(CameraDevice cameraDevice)
+        {
+            owner.mCameraOpenCloseLock.Release();
+            cameraDevice.Close();
+            owner.mCameraDevice = null;
+        }
+
+        public override void OnError(CameraDevice cameraDevice, CameraError error)
+        {
+            owner.mCameraOpenCloseLock.Release();
+            cameraDevice.Close();
+            owner.mCameraDevice = null;
+            if (owner == null)
+                return;
+            Activity activity = owner.Activity;
+            if (activity != null)
+            {
+                activity.Finish();
+            }
+
+        }
+    }
+}
